// enums/block-name.ts
export enum
BlockName {
  ACCORDION = 'Accordion',
  ALERT = 'Alert',
  BUTTON = 'Button',
  CALL_TO_ACTION = 'CallToAction',
  CARD = 'Card',
  CARD_CONTENT = 'Card.Content',
  CARD_HEADER = 'Card.Header',
  CARD_FOOTER = 'Card.Footer',
  CONTAINER = 'Container',
  DIVIDER = 'Divider',
  FORM = 'Form',
  ICON = 'Icon',
  IFRAME = 'Iframe',
  IMAGE = 'Image',
  INPUT = 'Input',
  LINK = 'Link',
  MARKDOWN = 'Markdown',
  RICH_TEXT = 'RichText',
  SELECT = 'Select',
  TEXT = 'Text',
  TEXTAREA = 'Textarea',
  TOGGLE = 'Toggle',
  ADD_TO_CALENDAR_BUTTON = 'AddToCalendarButton',
  ANNOUNCEMENT = 'Announcement',
  ANNOUNCEMENT_BANNER = 'AnnouncementBanner',
  HERO_BANNER = 'HeroBanner',
  LINK_MENU = 'LinkMenu',
  HTML_SCRIPT = 'HtmlScript'
}